Moved images from source folder to The Inversion folder, prefer them being there and may be the cause of the last commit not working even though all files were pushed!
Added tutorial image, back button sprite sheet, invert button and boost button
images.

Added the invert and boost buttons to the Game state, when pressed they activate
there relevant player effects.

Encapsulated the previous "player" sprite and effects into the RunnerPC "class".
Created the game Boost mechanic function for this RunnerPC class does as written
in the spec increases the players velocity by a large amount whilst stopping the
players downward velocity and any gravity applying to it.

Tweaked the gravity functionallity to include the rotation and flipping of the
sprite so that it's running in the correct way and not running on its head.

Created the score mechanic and added on screen text to display the values.

Created a Tutorial state that can now be switched to via the tutorial button
it has a back button to go back to the main menu and an tutorial image
describing how the game works.