Removed the html headers etc at the top of the Index (as they aren't needed), updated Phaser to v2.0.2 it fixes a few bugs (not any that I've had trouble with but can't be too careful).
Changed the position of the back button in Tutorial.

Added new bool parameter and set function to Runner PC it's a boolean that gets
set to true if the RunnerPC is touching the ground(Externally in Game) and false
if it isn't touching the ground. If it's true I.E touching something we can inve
-rt gravity if not we can't.

found a new set of open source images and I've added the edited versions of them
I made. Will add the text file with source etc in tne next push.
Preloaded them in, during the Init state.

Added a GameOver variable to Game to set the game to false when we've collided
with an obstacle. ceases the update loop.

Created WorldObject object which essentially creates an object in the world that
physics can not enact on similar to RunnerPC however it does not have any funct-
ons of it's own at the moment, so I'm unsure if there's a need in it overall.
However Obstacles & Platforms are made from this Object.

Added more game logic to the Game states update function as well as creating
Phaser Groups of obstacles (one for floor & roof tiles, one for platforms and
the last for obstacles) The floor and roof tiles are currently positioned proper
-ly but the Platforms and Obstacles are just in test positions. Essentially the
logic in update allows platforms to be walked on and updates the RunnerPC onFloo
-r bool when touching them. And the Obstacles incur game over when touched.

Also created Camera Lock on to the Player and an "infinite world" via looping of
the 3200x640 world around everytime the camera reaches the end, when it reaches
the end it negates all of the objects 900(camera width) from the 3200 ends x
value by 2300 effectively putting them at the start of the world but creating a
"Seamless" Infinite runner it also kills anything already between 0-900.

In future the plan is to randomize all the platforms and obstacles in this world
As at the moment its a bit sparse.