Not a lot of new additions to this commit in terms of mechanics, however I completely changed the way scrolling works, I changed it too the much more sens- ible everything scrolls towards the players character and the character remains still! This forced me to change a few dozen things in the player such as holding a velocity that doesn't get added to its own physics body but instead the invers -e gets added on to every other thing. I also had to change how the floor platfo -rms acted they now have to loop around once out of the screen(instead of the pl -ayer like last time).
Added randomization to the platforms and obstacles to position them. Spikes only
for obstacles at the moment, I intend to add another type of obstacle in later.
The current randomization is ok but could do with some more tweaking.

I also changed the WorldObjects to accept a boolean that states if its collideab
-le from the left side or not.

Added a scrolling background to the game, which was one of the main reasons i
changed to the current way the game scrolls! To simplify it a whole lot.

Added an animated sprite in for the player character instead of just an normal i
-mage, added several other paramaters to the RunnerPC constructor to make it mor
-e flexible for this.

Tweaked player velocity to always move at 50 when its on a platform (opposite of
its fake velocity that all the other world objects move at) this is to stop the
character being dragged to the other end of the screen instead of remaining cent
-ered like it should. As despite us wanting the player to remain still the platf
-orms moving in the opposite direction drag the player to the left whenever we
wish to land on them, due to phaser physics! And we can't turn the effects off
sadly, so got to use this slightly hacky alternative for now, might find a bett
-er way in the future.

Other than the above I done some other minor changes to support the scrolling
change, the code at the moment is still a bit messy and the commenting needs
updated a little! Will fix the majority of it on the next push.