A commit after a long break period from the project as such it contains a large number of new additions as I fell behind slightly in one task.
The Major changes in this commit are the inclusion of GameOverScoreStates
which control the internal game over menu that happens when you hit an obstacle
while playing. Alongside all the various actions you can take in these mini-menu
-s.

A Highscore object an object that displays and allows the transformation of a
string in real time based on user input, Highscores state(for showing the Highsc
-ore table, it currently works with a fixed high score table, have 2 functions
to aid in changing have yet to test them but logically they're sound and haven't
been invoked anywhere yet). Highscore Table basic table that renders the highsco
-res to screen, has add highscore function and check if the score is a highscore
functions for the time being, currently untested but should work, not in use at
the moment so they don't hinder the usabillity of the commit.

A few self made images relating to the GameOverStates and Highscores they're
largelly buttons.

Got the state transitions to display the previous game in the background as show
-n in my spec, had to make positions global to all states for this, not exactly
happy about it but, I tried some alternatives that didn't work and Phaser doesn
't handle passing objects between states it deletes them all on state switch. H-
opefully I'll find an alternative to this soon.

Various tweaks here and there to fix code, still needs a severe clean up, will
hopefully have time once finished.

Blog post will follow this commit tomorrow (pulled a 24 hour shift doing this
partially out of enjoyment and partially out of duty, so sleep comes first).