Multiple Error fixes to various aspects of the Highscore Table and how it inter- acts with the GameOverScoreState, Init and GameState. The highscore table is currently global to all states which I'm not a fan of however the scores need loaded in on initialization from the server, we need access to it in GameOverSc- oreState and we use it in the HighscoreTable so it's difficult to integrate it i -n a way that works out in all cases. Something i hope to go back to in my clean up over the next week.
Other than that this push contains the working Ajax HighscorreTable functions wh
-ich retrieve all the highscore functions from the server and push a single high
-score to the server. It's also setup to send the packets to the google app serv
-ver and retrieve them all at once as well and then load them into the Table so
that they're displayable on entry to the Highscore state.

I've also committed the server code even though its in Python and the repo is
javascript. Essentially its a modified version of the one in the Lab setup to ha
-ndle the highscores and only take in 10 unique key values relating to there pos
-ition on the highscore table in reference to there index value. I.E
1st = 0 2nd = 1 etc. It's currently fully working and works with an online googl
-e app server.

That's all for now as of just now the project is hand-inable however I really
wish to work on the gameplay code significantly over the next few days as it's
rather meh...right now.

I'll be writing my blog post to follow this up tomorrow after I've modified the
player physics to be more enjoyable.